+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/api/NetIdService.kt
@@ -14,17 +14,10 @@
 
 package de.netid.mobile.sdk.api
 
-import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.util.Log
-import android.view.View
-import android.widget.Button
-import android.widget.LinearLayout
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
-import com.google.android.material.button.MaterialButton
-import de.netid.mobile.sdk.R
 import de.netid.mobile.sdk.appauth.AppAuthManager
 import de.netid.mobile.sdk.appauth.AppAuthManagerFactory
 import de.netid.mobile.sdk.appauth.AppAuthManagerListener
@@ -38,6 +31,19 @@ import de.netid.mobile.sdk.util.JsonUtil
 import de.netid.mobile.sdk.util.PackageUtil
 import de.netid.mobile.sdk.util.ReachabilityUtil
 
+/**
+ * The ``NetIdService`` is the main class of the sdk.
+ * An application communicates via this class with the authorization service.
+ *
+ * To do so, an application first registers itself as a listener to the service.
+ * ``NetIdService.addListener(this)``
+ *
+ * Next, initialize the service with a configuration object of kind ``NetIdConfig``.
+ *
+ * The application has to conform to the ``NetIdServiceDelegate`` protocol and implement the required functions (see below).
+ *
+ * ``NetIdService.initialize(netIdConfig, this)``
+ */
 object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
     UserInfoManagerListener, PermissionManagerListener {
 
@@ -45,11 +51,6 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
     private const val broker = "broker.netid.de"
     private var layerStyle: NetIdLayerStyle = NetIdLayerStyle.Solid
     private var buttonStyle: NetIdButtonStyle = NetIdButtonStyle.WhiteSolid
-    private var netIdLogoResource:Int = R.drawable.ic_netid_logo_button
-    private var buttonBackgroundResource:Int = R.color.authorization_agree_button_color
-    private var buttonForegroundResource:Int = R.color.authorization_agree_text_color
-    private var buttonOutlineResource:Int = R.color.authorization_close_button_color
-    private var buttonStrokeWidthResource:Int = R.dimen.authorization_close_button_stroke_width
 
     private var netIdConfig: NetIdConfig? = null
 
@@ -66,14 +67,27 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
     private var appButtonFragmentsForLogin = mutableMapOf<String, Fragment>()
     private var appButtonFragmentsForLoginPermission = mutableMapOf<String, Fragment>()
 
+    /**
+     * Registers a new listener of type NetIdServiceListener
+     * @param listener The new listener.
+     */
     fun addListener(listener: NetIdServiceListener) {
         netIdServiceListeners.add(listener)
     }
 
+    /**
+     * Removes a listener of type NetIdServiceListener from the list of registered listeners.
+     * @param listener The listener to remove.
+     */
     fun removeListener(listener: NetIdServiceListener) {
         netIdServiceListeners.remove(listener)
     }
 
+    /**
+     * Initializes the sdk and loads the authentication configuration document.
+     * @param  netIdConfig The client configuration of type ``NetIdConfig``
+     * @param  context Context to use.
+     */
     fun initialize(netIdConfig: NetIdConfig, context: Context) {
         if (handleConnection(context, NetIdErrorProcess.Configuration)) {
             if (this.netIdConfig != null) {
@@ -152,7 +166,7 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
      * Sets the style to use for all buttons when using the button flow.
      * @param buttonStyle button style to set, can be any of ``NetIdButtonStyle``, defaults to ``NetIdButtonStyle.GraySolid``
      */
-    fun setButtonStyle(buttonStyle: NetIdButtonStyle, activity: Activity) {
+    fun setButtonStyle(buttonStyle: NetIdButtonStyle) {
         this.buttonStyle = buttonStyle
 
         if (permissionContinueButtonFragment != null) {
@@ -216,7 +230,7 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
      * @param context Context to use.
      * @return Authorization intent.
      */
-    fun authIntentForFlow(flow: NetIdAuthFlow, context: Context): Intent? {
+    internal fun authIntentForFlow(flow: NetIdAuthFlow, context: Context): Intent? {
         netIdConfig?.let { config ->
             return appAuthManager.getAuthorizationIntent(
                 config.clientId,
@@ -263,7 +277,8 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
      * @param key Key denoting one of the installed account provider apps. Use ``getKeysForAccountProviderApps`` first to get the keys/names of all installed account provider apps.
      * @param authFlow Can be any of .Permission, .Login or .LoginPermission.
      * @param continueText Alternative text to set on the button. If empty, the default will be used.
-     * @returns Button with text and label for the chosen id app. If index is out of bounds or no app is installed, returns an empty view.
+     * @returns Button with text and label for the chosen id app. If index is out of bounds or no app is installed, ArrayIndexOutOfBoundsException is thrown.
+     * @throws ArrayIndexOutOfBoundsException
      */
     fun accountProviderAppButtonFragment(key: String, flow: NetIdAuthFlow, continueText: String = ""): Fragment {
         val keys = getKeysForAccountProviderApps()