+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/ui/LoginContinueButtonFragment.kt
@@ -72,10 +72,10 @@ class LoginContinueButtonFragment(
         var netIdLogoResource = R.drawable.ic_netid_logo_small
         var buttonBackgroundResource = R.color.authorization_agree_button_color
         var buttonForegroundResource = R.color.authorization_agree_text_color
-        var buttonOutlineResource = R.color.authorization_close_button_color
+        var buttonOutlineResource = R.color.authorization_agree_outline_color
         var buttonStrokeWidthResource = R.dimen.authorization_close_button_stroke_width
 
-        when (NetIdService.getButtonStyle()) {
+        when (buttonStyle) {
             NetIdButtonStyle.GreenSolid -> {
                 netIdLogoResource = R.drawable.ic_netid_logo_button_white
                 buttonBackgroundResource = R.color.green_background_color
@@ -94,15 +94,15 @@ class LoginContinueButtonFragment(
                 netIdLogoResource = R.drawable.ic_netid_logo_small
                 buttonBackgroundResource = R.color.authorization_agree_button_color
                 buttonForegroundResource = R.color.authorization_agree_text_color
-                buttonOutlineResource = R.color.authorization_close_button_color
+                buttonOutlineResource = R.color.authorization_agree_outline_color
                 buttonStrokeWidthResource = R.dimen.authorization_close_button_stroke_width
             }
         }
 
-        binding.buttonLoginContinue.setTextColor(resources.getColor(buttonForegroundResource))
-        binding.buttonLoginContinue.setBackgroundColor(resources.getColor(buttonBackgroundResource))
+        binding.buttonLoginContinue.setTextColor(resources.getColor(buttonForegroundResource, null))
+        binding.buttonLoginContinue.setBackgroundColor(resources.getColor(buttonBackgroundResource, null))
         binding.buttonLoginContinue.setStrokeColorResource(buttonOutlineResource)
-        binding.buttonLoginContinue.icon = resources.getDrawable(netIdLogoResource)
+        binding.buttonLoginContinue.icon = resources.getDrawable(netIdLogoResource, null)
         binding.buttonLoginContinue.setStrokeWidthResource(buttonStrokeWidthResource)
     }