+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/userinfo/UserInfoManagerListener.kt
@@ -17,7 +17,7 @@ package de.netid.mobile.sdk.userinfo
 import de.netid.mobile.sdk.api.NetIdError
 import de.netid.mobile.sdk.model.UserInfo
 
-interface UserInfoManagerListener {
+internal interface UserInfoManagerListener {
 
     fun onUserInfoFetched(userInfo: UserInfo)