+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/util/TokenUtil.kt
@@ -19,7 +19,7 @@ import org.json.JSONException
 import org.json.JSONObject
 import java.nio.charset.StandardCharsets
 
-class TokenUtil {
+internal class TokenUtil {
 
     companion object {
         private const val claimPermissionManagement = "permission_management"